fine tuned getVps to get precise vertical value

+++ android/src/com/google/zxing/client/android/history/HistoryManager.java
@@ -161,7 +161,7 @@ public final class HistoryManager {
     }
   }
 
-  public void addHistoryItem(double magAxis[], double fusAxis[], float gpsAxis[], float orientation[], double distance, Result result, ResultHandler handler) {
+  public void addHistoryItem(double magAxis[], double fusAxis[], double gpsAxis[], float orientation[], double distance, Result result, ResultHandler handler) {
 	  // Do not save this item to the history if the preference is turned off, or the contents are
 	  // considered secure.
 	  if (!activity.getIntent().getBooleanExtra(Intents.Scan.SAVE_HISTORY, true) ||
